requirements.txt
appdirs==1.4.3
bcrypt==3.1.3
beautifulsoup4==4.5.3
cffi==1.9.1
cornice==2.4.0
decorator==4.0.11
hupper==0.4.3
Jinja2==2.9.5
MarkupSafe==1.0
olefile==0.44
packaging==16.8
PasteDeploy==1.5.2
Pillow==4.1.1
psycopg2-binary
py==1.4.32
pycparser==2.17
PyJWT==1.4.2
pyparsing==2.2.0
pyramid==1.8.1
pyramid-jinja2
pyramid-jwt
pyramid-tm
pytest==3.0.6
reportlab==3.4.0
repoze.lru==0.6
robotframework==3.0.2
robotframework-ride==1.5.2.1
robotframework-selenium2library==1.8.0
selenium==3.4.1
simplejson==3.10.0
six==1.10.0
SQLAlchemy==1.1.5
translationstring==1.3
venusian==1.0
waitress==1.0.2
WebOb==1.7.1
WebTest==2.0.26
zope.deprecation==4.2.0
zope.interface==4.3.3
gunicorn
zope.sqlalchemy
pyramid-mailer